Added IDReferences and Extrinsics traits where needed by DTD
